SBS-IAM-INTEGRATION
Each service will now communicate to IAM, the resource ID and action ID as two additional parameters in token validation API (new Authorize API). IAM will verify based on its policies and respond to the service with all the details that token API did, with an additional “Yes” or “No” response- whether a user is allowed to perform the action on that specific resource.

To cater to these new IAM changes, a new middleware is added; it communicates to IAM, the action and the associated resource corresponding to each API.

Note that this middleware is added for cinder-v2 API only. Since we don't support v1 we should disable v1 in prod through config options.

* cinder/policy.py
  + Disabled policy check at cinder

* etc/cinder/api-paste.ini
  + Added new middleware for cinder-v2 API

* cinder/auth_middleware/mapping.py
  + Mapping between API urls, actions and resources

* cinder/auth_middleware/__init__.py
  + Parse the request, figure out the action and associated resource and communicate it to IAM as enviroment variables
  + If the action/resource could not be looked up from the mapping, then its an unsupported API; return '404 Not Found'

Implements bp sbs-iam-integration