JBS-192: Update 1st snapshot timestamp to IST
We update 1st snapshot created_at time, so that it is higher
than the base snapshot.

With changes in timestamp to IST, the 1st snapshot was getting
UTC, which would be behind IST, leading to failure in deleting
that snapshot.

And further snapshots would be rejected on the volume, as the latest
snapshot would be stuck in deletion.

Signed-off-by: shishir gowda <shishir.gowda@ril.com>