JBS-190: Check if parent snapshot is available before creating new
Currently, we do not check if a snapshot is in creating state, and
accept a new snapshot create request on the same volume. This happens
as we reset volume status from backing up after src snapshot is done.

New snapshot entry gets created, but snapshot fails if previous snap
is still in progress, as entry in DSS would not be available.

The check is to make sure parent snapshot is in "available" state
 in the api layer itself.

Signed-off-by: shishir gowda <shishir.gowda@ril.com>